* Switch to using double precision (double) over single precision (float). This was done because Hercules & bluecrystal have cards that support   compute level >=1.3 . This means they can support double precision floating point numbers.

+++ differentiate.c
@@ -9,7 +9,7 @@
 #include "differentiate.h"
 
 	//Calculate partial derivative of Nx w.r.t to x using forward differencing at point (xPos,yPos) on LatticeObject l
-	float dNxdx_F(const LatticeObject* l, int xPos, int yPos)
+	double dNxdx_F(const LatticeObject* l, int xPos, int yPos)
 	{
 		/*see if we need to flip directorElement to make sure angle between
 		* vectors is < 90deg so if cos(theta) < 0 then we should flip directorElement by 180degress
@@ -29,7 +29,7 @@
 	}
 
 	//Calculate partial derivative of Nx w.r.t to x using backward differencing at point (xPos,yPos) on LatticeObject l
-	float dNxdx_B(const LatticeObject* l, int xPos, int yPos)
+	double dNxdx_B(const LatticeObject* l, int xPos, int yPos)
 	{
 		/*see if we need to flip directorElement to make sure angle between
 		* vectors is < 90deg so if cos(theta) < 0 then we should flip directorElement by 180degress
@@ -49,7 +49,7 @@
 	}
 
 	//Calculate partial derivative of Nx w.r.t to y using forward differencing at point (xPos,yPos) on LatticeObject l
-	float dNxdy_F(const LatticeObject* l, int xPos, int yPos)
+	double dNxdy_F(const LatticeObject* l, int xPos, int yPos)
 	{
 		/*see if we need to flip directorElement to make sure angle between
 		* vectors is < 90deg so if cos(theta) < 0 then we should flip directorElement by 180degress
@@ -69,7 +69,7 @@
 	}
 
 	//Calculate partial derivative of Nx w.r.t to y using backward differencing at point (xPos,yPos) on LatticeObject l
-	float dNxdy_B(const LatticeObject* l, int xPos, int yPos)
+	double dNxdy_B(const LatticeObject* l, int xPos, int yPos)
 	{
 		/*see if we need to flip directorElement to make sure angle between
 		* vectors is < 90deg so if cos(theta) < 0 then we should flip directorElement by 180degress
@@ -89,7 +89,7 @@
 	}
 
 	//Calculate partial derivative of Ny w.r.t to x using forward differencing at point (xPos,yPos) on LatticeObject l
-	float dNydx_F(const LatticeObject* l, int xPos, int yPos)
+	double dNydx_F(const LatticeObject* l, int xPos, int yPos)
 	{
 		/*see if we need to flip directorElement to make sure angle between
 		* vectors is < 90deg so if cos(theta) < 0 then we should flip directorElement by 180degress
@@ -109,7 +109,7 @@
 	}
 
 	//Calculate partial derivative of Ny w.r.t to x using backward differencing at point (xPos,yPos) on LatticeObject l
-	float dNydx_B(const LatticeObject* l, int xPos, int yPos)
+	double dNydx_B(const LatticeObject* l, int xPos, int yPos)
 	{
 		/*see if we need to flip directorElement to make sure angle between
 		* vectors is < 90deg so if cos(theta) < 0 then we should flip directorElement by 180degress
@@ -129,7 +129,7 @@
 	}
 
 	//Calculate partial derivative of Ny w.r.t to y using forward differencing at point (xPos,yPos) on LatticeObject l
-	float dNydy_F(const LatticeObject* l, int xPos, int yPos)
+	double dNydy_F(const LatticeObject* l, int xPos, int yPos)
 	{
 		/*see if we need to flip directorElement to make sure angle between
 		* vectors is < 90deg so if cos(theta) < 0 then we should flip directorElement by 180degress
@@ -149,7 +149,7 @@
 	}
 
 	//Calculate partial derivative of Ny w.r.t to y using backward differencing at point (xPos,yPos) on LatticeObject l
-	float dNydy_B(const LatticeObject* l, int xPos, int yPos)
+	double dNydy_B(const LatticeObject* l, int xPos, int yPos)
 	{
 		/*see if we need to flip directorElement to make sure angle between
 		* vectors is < 90deg so if cos(theta) < 0 then we should flip directorElement by 180degress
